pyproject.toml
[project]
name = "copernicus-downloader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.40.21",
    "cdsapi>=0.7.6",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
copernicus-downloader = "copernicus_downloader.main:main"
